--- api/hasura/src/__tests__/utils/date.ts
@@ -1,7 +1,0 @@
-function toEpoch(time: string) {
-  return new Date(time).getTime();
-}
-
-export default {
-  toEpoch,
-};
